% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{Exceedance}
\alias{Exceedance}
\title{Exceedance}
\usage{
Exceedance(measurements, OEL)
}
\arguments{
\item{measurements}{measurements of the SEG under assessment}

\item{OEL}{Occupational Exposure Limit of the agent}
}
\value{
probability of Exceedance
}
\description{
Probability of the OEL exceedance of the SEG considered
}
