% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{Ind.Compl}
\alias{Ind.Compl}
\title{Phase 3 in BOHS/NVVA 2018}
\usage{
Ind.Compl(x, Worker, Agent)
}
\arguments{
\item{x}{measurements and repeats of the SEG under assessment}
}
\value{
BW < 0.2totalVariance, True or False
}
\description{
Strategy of individual compliance performed in BOHS/NVVA 2018
}
