% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{Individual_Compliance}
\alias{Individual_Compliance}
\title{Phase 3, BOHS/NVVA 2011 - Individual Compliance}
\usage{
Individual_Compliance(seg, workers, samples, OEL)
}
\arguments{
\item{seg}{measurements with repeats of the SEG under assessment}

\item{workers}{workers code/name}

\item{samples}{samples concentrations of the agent}

\item{OEL}{Occupational Exposure Limit}
}
\value{
BW < 0.2totalVariance, Compliant or Not Compliant
}
\description{
Individual Compliance is achieved when there is less than 20% probability that
workers in a SEG have more than 5% of exposure greater than the OEL.
}
