% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{OverExp}
\alias{OverExp}
\title{OverExposure}
\usage{
OverExp(x, Worker, Agent, OEL)
}
\arguments{
\item{x}{measurements and repeats of the SEG under assessment}
}
\value{
% of OverExposure
}
\description{
OverExposure Calculation
}
