% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6891995.R
\name{phase1EN1995_k}
\alias{phase1EN1995_k}
\title{Phase 1, EN689 1995  - testing compliance for at least 3 measurements}
\usage{
phase1EN1995_k(samples, OEL)
}
\arguments{
\item{samples}{At least 3 measurements}

\item{OEL}{Occupational Exposure Limit of the agent}
}
\value{
Compliance (C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Compliance, uncertain compliance or non-compliance can be evaluated
for at least three measurements. Calculation of exposure indeces is performed (worker measurement/OEL).
Compliance is achieved if all indeces are below or equal to 0.25.
If at least an index is greater than 0.25 but all indeces are lower/equal than 1 and the
geometric mean of the measurements is lower/equal than 0.5, Compliance is accomplished.
Contrarly, there is no decision ("UC").
There is Non-Compliance if at least a measurement is greater than the OEL
and/or if at least an exposure index is greater than 1.
}
