% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6891995.R
\name{phase1EN1995_k1}
\alias{phase1EN1995_k1}
\title{Phase 1, EN689 1995  - testing compliance for 1 measurement}
\usage{
phase1EN1995_k1(measurements, OEL)
}
\arguments{
\item{measurements}{One measurement under assessment}

\item{OEL}{Occupational Exposure Limit of the agent}
}
\value{
Compliance ("C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Compliance, uncertain compliance or non-compliance can be evaluated
for one measurement. Calculation of exposure index is performed (worker measurement/OEL).
Compliance is achieved if the index is below or equal to 0.1.
If the index is greater than 0.1 but lower/equal than 1, there is no decision ("UC")
and additional measurements are necessary (at least three). Non-Compliance is accomplished
if the measurement is greater than the OEL and/or if the exposure index is greater than 1.
}
