% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6891995.R
\name{phase1EN1995_k3}
\alias{phase1EN1995_k3}
\title{Phase 1, EN689 1995  - testing compliance for 3 measurements}
\usage{
phase1EN1995_k3(x, OEL)
}
\arguments{
\item{x}{3 measurements (numeric)}

\item{OEL}{Establish the OEL (numeric)}
}
\value{
Compliance (C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Potential compliance, uncertain compliance or non-compliance of a SEG can be evaluated for three measurements. Index calculation is performed (worker measurement/OEL). Compliance is achieved if all of the index of the SEG are below or equal to 0.25 and below the OEL. There is Uncertain Compliance if at least one index is greater than 0.25 but below the OEL. There is Non-Compliance if at least an index is greater than OEL.
}
