% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6892018.R
\name{phase1EN2018_k3}
\alias{phase1EN2018_k3}
\title{Phase 1 for 1 measurement in EN689 2018}
\usage{
phase1EN2018_k3(samples)
}
\arguments{
\item{samples}{Three measurements under assessment}
}
\value{
Compliance ("C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Potential compliance, uncertain compliance or non-compliance of a SEG can be evaluated for three measurements. Index calculation is performed (worker measurement/OEL). Compliance is achieved if all of the index of the SEG are below or equal to 0.1 and below the OEL. There is Uncertain Compliance if at least one index is greater than 0.1 but below the OEL. There is Non-Compliance if at least an index is greater than OEL.
}
