% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseTwo_EN6892018.R
\name{phase2_UTL}
\alias{phase2_UTL}
\title{Phase 2 in EN689 2018}
\usage{
phase2_UTL(x)
}
\arguments{
\item{x}{measurements of the SEG under assessment}
}
\value{
UTL > OEL? TRUE or FALSE
}
\description{
Strategy of compliance performed in EN689 2018
}
