% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseTwo_EN6891995.R
\name{phase2EN689.1995}
\alias{phase2EN689.1995}
\title{Phase 2 in EN689 1995}
\usage{
phase2EN689.1995(x)
}
\arguments{
\item{x}{measurements of the SEG under assessment}
}
\value{
Green Area, Orange Area or Red Area
}
\description{
Strategy of compliance performed in EN689 1995
}
