% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{phase3_BoHS.NvVA}
\alias{phase3_BoHS.NvVA}
\title{Phase 3, BOHS/NvVA 2011  - Between Worker Variance}
\usage{
phase3_BoHS.NvVA(seg, workers, measurements)
}
\arguments{
\item{seg}{data of SEG under assessment}

\item{workers}{workers codes/names/ID}

\item{measurements}{samples concentrations of the agent}
}
\value{
BW < 0.2totalVariance ("Compliant"), BW > 0.2totalVariance ("Not Compliant")
}
\description{
Linear Mixed models using REML (Restricted Maximum Likelihood) is performed to calculate between worker variance
and total variance. The Compliance is achieved if the between worker variance is lower
than 0.2 of the total variance. Contrarily, there is Non-Compliance.
}
