% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{Exceedance}
\alias{Exceedance}
\title{Exceedance}
\usage{
Exceedance(samples, OEL)
}
\arguments{
\item{samples}{exposure concentrations of the SEG under assessment}
\item{OEL}{Occupational Exposure Limit}
}
\value{
% of Exceedance
}
\description{
Probability exceedance of workers exposures/SEGs in log-normal distribution
}
