% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{Ind.Compl}
\alias{Ind.Compl}
\title{Individual Compliance in BOHS/NVvA - 2011}
\usage{
Ind.Compl(x, Workers, Agent)
}
\arguments{
\item{x}{Similar exposed group- SEG}
\item{Workers}{Workers ID}
\item{Agent}{measurements with at least one repeat}
}
\value{numeric outputs}

\description{
{Individual compliance is achieved when more than 80 percent of SEG members have at least 95 percent of their exposures < OEL. Contrarily, there is non-compliance}
}
