% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{Ind.Compl}
\alias{Ind.Compl}
\title{Individual Compliance in BOHS/NVvA - 2011}
\usage{
Ind.Compl(seg, workers, samples, OEL)
}
\arguments{
\item{seg}{Similarly Exposed Group}
\item{workers}{Workers ID}
\item{samples}{Measurements with at least one repeat}
\item{OEL}{Occupational Exposure Limit}
}
\value{numeric outputs}

\description{
{Individual compliance is achieved when more than 80 percent of SEG members have at least 95 percent of their exposures < OEL. Contrarily, there is non-compliance}
}
