% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{OverExp}
\alias{OverExp}
\title{OverExposure}
\usage{
OverExp(seg, workers, samples, OEL)
}
\arguments{
\item{seg}{measurements (with repeated measurements) of the SEG under assessment}

\item{workers}{workers codes/ID}

\item{samples}{exposure concentrations of workers}

\item{OEL}{Occupational Exposure Limit of the agent}
}
\value{
% of OverExposure
}
\description{
OverExposure Calculation
}
