% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{OverExp}
\alias{OverExp}
\title{OverExposure}
\usage{
OverExp(seg, workers, samples, OEL)
}
\arguments{
\item{seg}{Similarly Exposed Group}
\item{workers}{Workers ID}
\item{samples}{Exposure concentrations}
\item{OEL}
}
\value{numeric outputs
% of OverExposure
}
\description{
Overexposure is the probability that
the mean exposures of a randomly selected worker can be greater
than the OEL.
}
