% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExp_Exc.R
\name{OverExp}
\alias{OverExp}
\title{OverExposure}
\usage{
OverExp(x, Workers, Agent, OEL)
}
\arguments{
\item{x}{SEG}
\item{Workers}{Workers ID}
\item{Agent}{measurements under-assessment}
\item{OEL}
}
\value{numeric outputs
% of OverExposure
}
\description{
Overexposure is the probability that
the mean exposure of a randomly selected worker is greater
than the OEL.
}
