% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6891995.R
\name{phase1EN1995_k}
\alias{phase1EN1995_k}
\title{Phase 1, EN689 1995  - testing compliance for 3 or > measurements}
\usage{
phase1EN1995_k(samples, OEL)
}
\arguments{
\item{samples}{3 or > measurements}

\item{OEL}{Occupational Exposure Limit}
}
\value{
Compliance ("C"), Ucertainty Compliance ("UC), Non-Compliance ("NC")
}
\description{
Potential compliance, uncertain compliance or non-compliance of a SEG can be evaluated for three or more measurements. Index calculation is performed (worker measurement/OEL). Compliance is achieved if all of the index of the SEG are below or equal to 1 and below the OEL, and the GM is below 0.5. There is Uncertain compliance if at least one index is greater than 1 or/and when the GM is greater than 0.5, but all the index are below the OEL. There is Non-Compliance if at least an index is greater than OEL.
}
