% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6891995.R
\name{phase1EN1995_k1}
\alias{phase1EN1995_k1}
\title{Phase 1, EN689 1995  - testing compliance for 1 measurement}
\usage{
phase1EN1995_k1(samples, OEL)
}
\arguments{
\item{samples}{One measurement under assessment}
\item{OEL}{Occupational Exposure Limit}
}
\value{
Compliance ("C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Compliance, Uncertain Compliance or Non-Compliance of a SEG can be evaluated for one measurement. Index calculation is performed (worker measurement/OEL). Compliance is achieved if the index is below or equal to 0.1 and below the OEL. There is Uncertain Compliance if the index is greater than 0.1 but below the OEL. There is Non-Compliance if the index is greater than OEL.
}
