% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6892018.R
\name{phase1EN2018_k3}
\alias{phase1EN2018_k3}
\title{Phase 1 for 1 measurement in EN689 2018}
\usage{
phase1EN2018_k3(x)
}
\arguments{
\item{x}{Three measurements under assessment}
}
\value{
Compliance ("C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Strategy of compliance performed in EN689 2018
}
