% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseOne_EN6892018.R
\name{phase1EN2018_k3}
\alias{phase1EN2018_k3}
\title{Phase 1, EN689 2018  - testing compliance for 3 measurements}
\usage{
phase1EN2018_k3(measurements, OEL)
}
\arguments{
\item{measurements}{Three measurements}

\item{OEL}{Occupational Exposure Limit of the agent}
}
\value{
Compliance ("C"), Non-Compliance ("NC") or Uncertainty of Compliance ("UC")
}
\description{
Compliance, uncertain compliance or non-compliance of a SEG can be
evaluated for three measurements. Compliance is achieved if all of
the measurements are below or equal to 0.1xOEL. There is Uncertain Compliance if
at least a measurement is greater than 0.1xOEL but below the OEL.
There is Non-Compliance if at least a measurement is greater than OEL.
}
