% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseTwo_EN6892018.R
\name{phase2_UTL}
\alias{phase2_UTL}
\title{Phase 2/Statistical test in EN689:2018}
\usage{
phase2_UTL(samples)
}
\arguments{
\item{samples}{exposure concentrations}
\item{OEL}{Occupational Exposure Limit}
}
\value{
UTL > OEL? TRUE or FALSE
}
\description{
Upper tolerance limit value with 95 percent confidence interval and 70 percent confidence level is the threshold probability parameter for OEL exceedance on the log-normal distribution.
}
