% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseTwo_EN6892018.R
\name{phase2_Uvalue}
\alias{phase2_Uvalue}
\title{Phase 2 in EN689 2018}
\usage{
phase2_Uvalue(x)
}
\arguments{
\item{x}{measurements of the SEG under assessment from 6 to 10}
}
\value{
U value > or < that U thresholds
}
\description{
Strategy of compliance performed in EN689 2018
}
