% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseTwo_EN6891995.R
\name{phase2EN689.1995}
\alias{phase2EN689.1995}
\title{Phase 2 in EN689 1995}
\usage{
phase2EN689.1995(samples)
}
\arguments{
\item{samples}{exposure concentrations}
\item{OEL}{Occupational Exposure Limit}
}
\value{
Green Area, Orange Area or Red Area
}
\description{
5 percent exc. is 5 percent probability for OEL exceedance on the values distribution. Green area: less or equal 1 percent of the measurements are higher than OEL. Orange area: less or equal 5 percent of the measurements are higher than OEL. Red Area: More than 5 percent of the measurements are higher than OEL
}
