% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{phase3_BoHS.NvVA}
\alias{phase3_BoHS.NvVA}
\title{Phase 3, BOHS/NvVA 2011  - Between Worker Variance}
\usage{
phase3_BoHS.NvVA(seg, workers, samples)
}
\arguments{
\item{seg}{measurements with repeats of the SEG under assessment}

\item{workers}{workers codes/names/ID}

\item{samples}{samples concentrations of the agent}
}
\value{
BW < 0.2totalVariance ("Compliant"), BW > 0.2totalVariance ("Not Compliant")
}
\description{
REML (Restricted Maximum Likelihood) is performed to calculate between worker variance
and total variance. The Compliance is achieved if the between worker variance is lower
than 0.2 of the total variance. Contrarly, there is Non-Compliance.
}
