% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{phase3BoHS.NvVA}
\alias{phase3BoHS.NvVA}
\title{Phase 3 in BOHS/NVVA 2018}
\usage{
phase3BoHS.NvVA(samples, workers, seg)
}
\arguments{
\item{seg}{Similarly Exposed Group}
\item{samples}{measurements with at least one repeat}
\item{workers}{Workers ID}
}
\value{
True or False
}
\description{
Between worker variance compliance is reached when it accounts to be 20 percent of the total variance
}
