% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseThree_BOHS.NVvA2018.R
\name{phase3BoHS.NvVA}
\alias{phase3BoHS.NvVA}
\title{Phase 3 in BOHS/NVVA 2018}
\usage{
phase3BoHS.NvVA(Agent, Worker, x)
}
\arguments{
\item{x}{measurements and repeats of the SEG under assessment}
}
\value{
BW < 0.2totalVariance, True or False
}
\description{
Strategy of compliance performed in BOHS/NVVA 2018
}
